requirements.txt
certifi==2021.5.30
pycrypto==2.6.1
pytz==2021.1
wincertstore==0.2
